requirements.txt
boto3==1.28.57
python-dateutil~=2.8.2